refactor: rename transaction building module
