+++ src/lock.ts
